backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pyyaml>=5.3
tqdm>=4.41.0
matplotlib>=3.2.2
seaborn>=0.11.0
requests>=2.23.0
pandas>=1.1.4
scipy>=1.4.1

